Moved ImportHelper from views to utils with minor changes.
